requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
psycopg2-binary==2.9.2
pydantic==1.8.2
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
typing-extensions==4.0.1
uvicorn==0.16.0
